Move files into new holidays package and update dependencies.
