+++ app/lib/blocs/dashbord_widgets_blocs/holiday_bloc.dart
@@ -7,13 +7,12 @@
 // SPDX-License-Identifier: EUPL-1.2
 
 import 'dart:async';
+
 import 'package:bloc_base/bloc_base.dart';
+import 'package:holidays/holidays.dart';
 import 'package:meta/meta.dart';
 import 'package:rxdart/subjects.dart';
-import 'package:sharezone/models/extern_apis/holiday.dart';
 import 'package:sharezone/util/api/user_api.dart';
-import 'package:sharezone/util/holidays/api_cache_manager.dart';
-import 'package:sharezone/util/holidays/state.dart';
 import 'package:user/user.dart';
 
 class HolidayBloc extends BlocBase {
