+++ app/lib/dashboard/dashboard_page.dart
@@ -13,6 +13,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_staggered_animations/flutter_staggered_animations.dart';
+import 'package:holidays/holidays.dart' hide State;
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:sharezone/blocs/application_bloc.dart';
 import 'package:sharezone/blocs/dashbord_widgets_blocs/holiday_bloc.dart';
@@ -24,7 +25,6 @@ import 'package:sharezone/dashboard/tips/dashboard_tip_system.dart';
 import 'package:sharezone/dashboard/update_reminder/update_reminder_bloc.dart';
 import 'package:sharezone/dashboard/widgets/blackboard_card_dashboard.dart';
 import 'package:sharezone/download_app_tip/widgets/download_app_tip_card.dart';
-import 'package:sharezone/models/extern_apis/holiday.dart';
 import 'package:sharezone/navigation/logic/navigation_bloc.dart';
 import 'package:sharezone/navigation/models/navigation_item.dart';
 import 'package:sharezone/navigation/scaffold/app_bar_configuration.dart';
