+++ app/test/dashboard/dashboard_page_test.dart
@@ -13,10 +13,10 @@ import 'package:bloc_provider/bloc_provider.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:golden_toolkit/golden_toolkit.dart';
+import 'package:holidays/holidays.dart';
 import 'package:intl/intl.dart';
 import 'package:sharezone/blocs/dashbord_widgets_blocs/holiday_bloc.dart';
 import 'package:sharezone/dashboard/dashboard_page.dart';
-import 'package:sharezone/models/extern_apis/holiday.dart';
 
 class FakeHolidayBloc extends Fake implements HolidayBloc {
   StreamController<bool> hasStateSelectedController = StreamController();
