+++ app/test/holidays/holiday_api_test.dart
@@ -8,11 +8,9 @@
 
 import 'dart:io';
 
+import 'package:holidays/holidays.dart';
 import 'package:http/http.dart' as http;
 import 'package:mockito/mockito.dart';
-import 'package:sharezone/models/extern_apis/holiday.dart';
-import 'package:sharezone/util/holidays/holiday_api.dart';
-import 'package:sharezone/util/holidays/state.dart';
 import 'package:test/test.dart';
 
 class HttpClientMock extends Mock implements http.Client {}
