+++ lib/holidays/lib/holidays.dart
@@ -1,0 +1,15 @@
+// Copyright (c) 2022 Sharezone UG (haftungsbeschränkt)
+// Licensed under the EUPL-1.2-or-later.
+//
+// You may obtain a copy of the Licence at:
+// https://joinup.ec.europa.eu/software/page/eupl
+//
+// SPDX-License-Identifier: EUPL-1.2
+
+library holidays;
+
+export 'src/api_cache_manager.dart';
+export 'src/api/holiday.dart';
+export 'src/state.dart';
+export 'src/holiday_api.dart';
+export 'src/holiday_cache.dart';
