+++ lib/holidays/lib/src/api/holiday.dart
@@ -13,7 +13,7 @@ import 'dart:convert';
 import 'package:built_collection/built_collection.dart';
 import 'package:built_value/built_value.dart';
 import 'package:built_value/serializer.dart';
-import 'package:sharezone/models/serializers.dart';
+import 'serializers.dart';
 
 part 'holiday.g.dart';
