+++ lib/holidays/pubspec.yaml
@@ -1,0 +1,24 @@
+name: holidays
+description: A starting point for Dart libraries or applications.
+version: 1.0.0
+
+publish_to: none
+
+environment:
+  sdk: ">=2.7.0 <3.0.0"
+
+dependencies:
+  built_collection: ^5.1.0
+  built_value: ^8.1.1
+  http: ^0.13.3
+  key_value_store:
+    path: ../key_value_store
+  cloud_functions: ^3.0.4
+  app_functions:
+    path: ../app_functions
+
+dev_dependencies:
+  build_runner: ^2.1.4
+  built_value_generator: ^8.1.1
+  lints: ^1.0.0
+  test: ^1.16.0
